implement delete all completed in javascript
